+++ cli/src/config.ts
@@ -1,7 +1,8 @@
 import { accessSync, readFileSync } from 'fs';
 import { basename, join, resolve } from 'path';
-import chalk from 'chalk';
+import prompts from 'prompts';
 
+import c from './colors';
 import { logFatal, readJSON } from './common';
 import { CliConfig, ExternalConfig, OS, PackageJson } from './definitions';
 
@@ -78,20 +79,8 @@ export class Config implements CliConfig {
     extConfig: ExtConfig,
     bundledWebRuntime: false,
     plugins: {},
-    assets: {
-      templateName: 'app-template',
-      templateDir: '',
-      pluginsTemplateDir: '',
-    },
     server: {
       cleartext: false,
-    },
-  };
-
-  plugins = {
-    assets: {
-      templateName: 'plugin-template',
-      templateDir: '',
     },
   };
 
@@ -101,7 +90,22 @@ export class Config implements CliConfig {
   constructor(os: string, currentWorkingDir: string, cliBinDir: string) {
     this.initOS(os);
     this.initCliConfig(cliBinDir);
-    this.setCurrentWorkingDir(currentWorkingDir);
+
+    try {
+      this.initAppConfig(resolve(currentWorkingDir));
+      this.loadExternalConfig();
+      this.mergeConfigData();
+
+      // Post-merge
+      this.initAndroidConfig();
+      this.initIosConfig();
+      this.initWindowsConfig();
+      this.initLinuxConfig();
+
+      this.knownPlatforms.push(this.web.name);
+    } catch (e) {
+      logFatal(`Unable to load config\n` + e.stack ?? e);
+    }
   }
 
   initOS(os: string) {
@@ -118,25 +122,6 @@ export class Config implements CliConfig {
     }
   }
 
-  setCurrentWorkingDir(currentWorkingDir: string) {
-    try {
-      this.initAppConfig(resolve(currentWorkingDir));
-      this.initPluginsConfig();
-      this.loadExternalConfig();
-      this.mergeConfigData();
-
-      // Post-merge
-      this.initAndroidConfig();
-      this.initIosConfig();
-      this.initWindowsConfig();
-      this.initLinuxConfig();
-
-      this.knownPlatforms.push(this.web.name);
-    } catch (e) {
-      logFatal(`Unable to load config`, e);
-    }
-  }
-
   private initCliConfig(cliBinDir: string) {
     this.cli.binDir = cliBinDir;
     this.cli.rootDir = join(cliBinDir, '../');
@@ -147,10 +132,6 @@ export class Config implements CliConfig {
   private initAppConfig(currentWorkingDir: string) {
     this.app.rootDir = currentWorkingDir;
     this.app.package = loadPackageJson(currentWorkingDir);
-    this.app.assets.templateDir = join(
-      this.cli.assetsDir,
-      this.app.assets.templateName,
-    );
   }
 
   async updateAppPackage() {
@@ -225,13 +206,6 @@ export class Config implements CliConfig {
     }
   }
 
-  private initPluginsConfig() {
-    this.plugins.assets.templateDir = join(
-      this.cli.assetsDir,
-      this.plugins.assets.templateName,
-    );
-  }
-
   private mergeConfigData() {
     const extConfig: ExternalConfig = this.app.extConfig || {};
 
@@ -251,7 +225,10 @@ export class Config implements CliConfig {
         // we've got an capacitor.json file, let's parse it
         this.app.extConfig = JSON.parse(extConfigStr);
       } catch (e) {
-        logFatal(`error parsing: ${basename(this.app.extConfigFilePath)}\n`, e);
+        logFatal(
+          `Error parsing ${basename(this.app.extConfigFilePath)}\n` + e.stack ??
+            e,
+        );
       }
     } catch {
       // it's ok if there's no capacitor.json file
@@ -268,7 +245,7 @@ export class Config implements CliConfig {
       const platformName = selectedPlatformName.toLowerCase().trim();
 
       if (!this.isValidPlatform(platformName)) {
-        logFatal(`Invalid platform: ${platformName}`);
+        logFatal(`Invalid platform: ${c.input(platformName)}`);
       } else if (!this.platformDirExists(platformName)) {
         this.platformNotCreatedError(platformName);
       }
@@ -287,25 +264,27 @@ export class Config implements CliConfig {
     promptMessage: string,
   ): Promise<string> {
     if (!selectedPlatformName) {
-      const inquirer = await import('inquirer');
+      const answers = await prompts(
+        [
+          {
+            type: 'select',
+            name: 'mode',
+            message: promptMessage,
+            choices: this.knownPlatforms.map(p => ({ title: p, value: p })),
+          },
+        ],
+        { onCancel: () => process.exit(1) },
+      );
 
-      const answer = await inquirer.prompt({
-        type: 'list',
-        name: 'mode',
-        message: promptMessage,
-        choices: this.knownPlatforms,
-      });
-
-      return answer.mode.toLowerCase().trim();
+      return answers.mode.toLowerCase().trim();
     }
 
     const platformName = selectedPlatformName.toLowerCase().trim();
 
     if (!this.isValidPlatform(platformName)) {
       logFatal(
-        `Invalid platform: "${platformName}". Valid platforms include: ${this.knownPlatforms.join(
-          ', ',
-        )}`,
+        `Invalid platform: ${c.input(platformName)}.\n` +
+          `Valid platforms include: ${this.knownPlatforms.join(', ')}`,
       );
     }
 
@@ -350,15 +329,15 @@ export class Config implements CliConfig {
   platformNotCreatedError(platformName: string) {
     if (platformName === 'web') {
       logFatal(
-        `Could not find the web platform directory. Make sure ${chalk.bold(
-          this.app.webDir,
-        )} exists.`,
+        `Could not find the web platform directory.\n` +
+          `Make sure ${c.strong(this.app.webDir)} exists.`,
       );
     }
     logFatal(
-      `${chalk.bold(
-        platformName,
-      )}" platform has not been created. Use "npx cap add ${platformName}" to add the platform project.`,
+      `${c.strong(platformName)} platform has not been added yet.\n` +
+        `Use ${c.input(
+          `npx cap add ${platformName}`,
+        )} to add the platform to your project.`,
     );
   }
 }