+++ cli/src/tasks/update.ts
@@ -1,3 +1,4 @@
+import c from '../colors';
 import { Config } from '../config';
 import { updateAndroid } from '../android/update';
 import { updateIOS, updateIOSChecks } from '../ios/update';
@@ -6,16 +7,12 @@ import {
   CheckFunction,
   check,
   checkPackage,
-  log,
-  logError,
   logFatal,
-  logInfo,
   resolvePlatform,
   runPlatformHook,
   runTask,
 } from '../common';
-
-import chalk from 'chalk';
+import { logger } from '../log';
 
 export async function updateCommand(
   config: Config,
@@ -27,14 +24,15 @@ export async function updateCommand(
     if (platformDir) {
       await runPlatformHook(platformDir, 'capacitor:update');
     } else {
-      logError(`platform ${selectedPlatformName} not found`);
+      logger.error(`Platform ${c.input(selectedPlatformName)} not found.`);
     }
   } else {
     const then = +new Date();
     const platforms = config.selectPlatforms(selectedPlatformName);
     if (platforms.length === 0) {
-      logInfo(
-        `There are no platforms to update yet. Create one with "capacitor create".`,
+      logger.info(
+        `There are no platforms to update yet.\n` +
+          `Add platforms with ${c.input('npx cap add')}.`,
       );
       return;
     }
@@ -48,9 +46,9 @@ export async function updateCommand(
       );
       const now = +new Date();
       const diff = (now - then) / 1000;
-      log(`Update finished in ${diff}s`);
+      logger.info(`Update finished in ${diff}s`);
     } catch (e) {
-      logFatal(e);
+      logFatal(e.stack ?? e);
     }
   }
 }
@@ -80,17 +78,14 @@ export async function update(
   deployment: boolean,
 ) {
   try {
-    await runTask(
-      chalk`{green {bold update}} {bold ${platformName}}`,
-      async () => {
-        if (platformName === config.ios.name) {
-          await updateIOS(config, deployment);
-        } else if (platformName === config.android.name) {
-          await updateAndroid(config);
-        }
-      },
-    );
+    await runTask(c.success(c.strong(`update ${platformName}`)), async () => {
+      if (platformName === config.ios.name) {
+        await updateIOS(config, deployment);
+      } else if (platformName === config.android.name) {
+        await updateAndroid(config);
+      }
+    });
   } catch (e) {
-    logError('Error running update:', e);
+    logger.error(`Error running update:\n` + e.stack ?? e);
   }
 }